tools: add nodetool script
This sript is intended as a wrapper for nodetool, first trying to route
commands to scylla-nodetool, and if the command is unimplemented, it is
routed to the java nodetool. Whether a command is implemented or not is
determined by checking the exit-code of scylla-nodetool. An
unimplemented command uses a unique exit-code of 100. When this is seen,
the command is routed to Java nodetool.
The script is designed to work both in-repo (when invoked as
./tools/nodetool) or at the final installed location too. In both cases,
it tries to come up with meaningful defaults for the location of the
scylla and nodetool executables respectively.
That said, the user can override these paths with the `SCYLLA_PATH` and
`NODETOOL_PATH` environment variables.

TODO: packaging

Refs: #15588