storage_service: Use storage_service::_sys_dist_ks in some places
The main goal here is to drop sys.dist.ks argument from the
init_messaging_service call to make future patching simpler. While doing
it it turned out that the argument was needed to be passed all the way
down to the mark_existing_views_as_built(), so this patch also dropes
this argument from this whole call-trace.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>