dist/redhat: remove tags for subpackage which are same as main preamble
this is a cleanup.

if a subpackage is licensed under a different license from the one
specified in the main preamble, we need to use a distinct License
tag on a per-subpackage basis. but if it is licensed with the
identical license, it is not necessary. since all three
subpackages of "*-{server, conf, kernel-conf}" are licensed under
AGPLv3, there is no need to repeat the "License:" tag in their
own preamble section.

the same applies to the "URL" tag.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>