storage_service: Dont capture container() on RPC handler
The handlers are about to be initialized from inside storage_service
constructor. At that time container() is not yet available and its
invalid to capture it on handlers' lambda. Fortunately, there's only one
handler that does it, other handlers capture 'this' and call container()
explicitly. This patch fixes the remaining one to do the same.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>