service/migration_manager: only reload schema when enabling disabled features
Instead of unconditionally reloading schema when enabling any schema
feature, only create a listener, if the feature was disabled in the
first place. So that we don't trigger reloading of the schema on each
schema feature, on node restarts. In this case, the node will start with
all these features enabled already.
This prevents unnecessary work on restarts.

Fixes: #16112

Closes scylladb/scylladb#16118