cql3: schema_altering_statement: pass `group0_guard` to `prepare_schema_mutations`
Until now `prepare_schema_mutations` would be passed only the write
timestamp from the guard.

In the next commit, I want to use some other information stored in the
guard in one of the implementations of `prepare_schema_mutations`
(namely, whether the schema change is performed with or without Raft).
Instead of extending `prepare_schema_mutations` with another parameter
(which most implementations won't use), replace the timestamp parameter
with a reference to the guard.