migration_manager: don't include base mutations when updating view schema in Raft mode
This logic of "including base mutations" when sending a view schema
change was implemented in case if a node receiving the view schema
mutations somehow missed earlier base mutations that the view mutations
may depend on.

First, in Raft mode it's not necessary to do this. The base mutations
are guaranteed to already be present on any node that applies the view
update, which is ordered after any base update that it may depend on in
the Raft log.

Second, including base mutations, due to the way it's implemented, may
cause problems. Suppose we have a base table with a view attached. When
we update the schema of this base table, the view's schema is also
updated. The base update generates mutations for base, obviously. But
the view update also generates mutations for base -- the "include base
mutations" logic.  It turns out that base mutations from base update
conflict with base mutations from the view update. This conflict is
currently benign.  But it manifested and caused a hard to debug failure
when we implemented group 0 schema versioning. It had to be reverted.

As a prerequisite to get group 0 schema versioning merged back, we need
to deal with this conflict. One way is to not include these base
mutations as part of view update (in Raft mode), which is what this
commit does.

Refs:
scylladb/scylladb#15331 (group 0 schema versioning PR)
scylladb/scylladb#15530 (issue caused by the conflicting mutations)
https://github.com/scylladb/scylladb/pull/15687#issuecomment-1755861099
(in this comment I describe in detail how the conflict works)