cql3: test_assignment: define formatter for assignment_testable
add fmt formatter for `assignment_testable`.

this is a part of a series to migrating from `operator<<(ostream&, ..)`
based formatting to fmtlib based formatting. the goal here is to enable
fmtlib to print `assignment_testabe` without the help of `operator<<`.

since we are still printing the shared_ptr<assignment_testable> using
operator<<(.., const assignment_testable&), we cannot drop this operator
yet.

Refs #13245
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>