commitlog: add formatters for commitlog::segment and friends
before this change, we rely on the default-generated fmt::formatter
created from operator<<, but fmt v10 dropped the default-generated
formatter.

in this change, we define formatters for following types:

- db::commitlog::segment
- db::commitlog::segment::cf_mark
- db::replay_position

since db::commitlog::segment's operator<< based formatter is only
used for providing fmtlib < v10, and it is implemented based on
format_as(), it is disabled when compiling with fmtlib >= 10. this
helps us to identify these helpers once we ditch the support of
fmtlib < 10, so we can drop them by grepping the FMT_VERSION macro.

since the existing generic formatter for std::vector<T> is still using
the operator<< based formatter, and it in turn calls the generic
operator<< overload for pair<U, V>. this does not work anymore before
we ditch the operator<< for std::vector<T> after this change, so
to accomodate the needs of formatting in both fmtlib v10 and fmtlib v9,
we have to use fmt::join() as an intermediate solution in this
intermediate phase.

Refs #13245
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>